1)Rename trustm_ec_key_helper to trustm_key_helper
