+++ trustm_provider/trustm_provider_signature.c
@@ -5,7 +5,7 @@
 #include <openssl/params.h>
 
 #include "trustm_provider_common.h"
-#include "trustm_ec_key_helper.h"
+#include "trustm_key_helper.h"
 
 typedef struct trustm_signature_ctx_str {
     const OSSL_CORE_HANDLE *core;
